backend/requirements.txt
# AI Book Splitter & Designer - Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.23.8
reportlab==4.0.7
pdfminer.six==20221105

# Document Processing
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
PyYAML==6.0.1
pandas==2.1.3
numpy==1.25.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# File Handling
aiofiles==23.2.1
pathlib2==2.3.7

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Image Processing (for PDF with images)
Pillow==10.1.0

# CORS
python-cors==1.7.0

# Environment Variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AI/ML Libraries (for advanced text analysis)
openai==1.3.7
# transformers==4.35.2
# torch==2.1.1

# Optional: Database (if needed in future)
# sqlalchemy==2.0.23
# alembic==1.12.1
# psycopg2-binary==2.9.9
